docs: stop using API routes because of the fair limits
